6.01
6.01.19
final version of test work

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -41,13 +53,33 @@ class PostsController extends Controller
     {
         $request->validate([
         'title'=> 'required',
-        'body' => 'required'
+        'body' => 'required',
+        'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File Upload
+
+        if($request->hasFile('cover_image')) {
+            // Get the filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename."_".time().".".$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.png';
+        }
 
         // Create Post
         $post = new Post;
         $post->body  = $request->input('body');
         $post->title = $request->input('title');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
 
         $post->save();
         return redirect('/posts')->with('success', 'Post has been added');
@@ -73,7 +105,13 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        //
+        $post = Post::find($id);
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+
+        return view('posts.edit')->with('post', $post);
     }
 
     /**
@@ -85,9 +123,31 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $request->validate([
+            'title'=> 'required',
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
+            ]);
+        if($request->hasFile('cover_image')) {
+            // Get the filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename."_".time().".".$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
+
       $post = Post::find($id);
       $post->title = $request->get('title');
       $post->body = $request->get('body');
+      if($request->hasFile('cover_image')){
+        $post->cover_image = $fileNameToStore;
+      }
       $post->save();
 
       return redirect('/posts')->with('success', 'Post has been updated');
@@ -102,8 +162,19 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.png'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
 
-        return redirect('/posts')->with('success', 'Stock has been deleted Successfully');
+        return redirect('/posts')->with('success', 'Post has been deleted Successfully');
     }
 }
